requirements.txt
alembic==1.2.1
asn1crypto==0.24.0
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
cffi==1.11.5
Click==7.0
colorlog==4.0.2
constantly==15.1.0
cryptography==2.4.2
Flask==1.1.1
Flask-API==1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
kraken-wsclient-py==0.0.4
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-dateutil==2.8.0
python-editor==1.0.4
service-identity==18.1.0
six==1.12.0
SQLAlchemy==1.3.10
Twisted==18.9.0
txaio==18.8.1
Werkzeug==0.16.0
zope.interface==4.6.0
